+++ tests/Support/Stubs/Forms/SearchUsersFormRequest.php
@@ -1,0 +1,28 @@
+<?php declare(strict_types=1);
+
+namespace Somnambulist\Bundles\FormRequestBundle\Tests\Support\Stubs\Forms;
+
+use Somnambulist\Bundles\FormRequestBundle\Http\FormRequest;
+
+/**
+ * Class SearchUsersFormRequest
+ *
+ * @package    Somnambulist\Bundles\FormRequestBundle\Tests\Support\Stubs\Forms
+ * @subpackage Somnambulist\Bundles\FormRequestBundle\Tests\Support\Stubs\Forms\SearchUsersFormRequest
+ */
+class SearchUsersFormRequest extends FormRequest
+{
+    protected array $ignore = ['page', 'per_page', 'include'];
+
+    public function rules(): array
+    {
+        return [
+            'name'  => 'nullable|string|max:100',
+            'email' => 'nullable|string|max:255',
+
+            'page'     => 'default:1|numeric',
+            'per_page' => 'default:30|numeric|max:100',
+            'include'  => ['nullable', 'regex:/[(groups|permissions),.]/'],
+        ];
+    }
+}